move user details to new seperate section
